refactor(commands): isolate admin mutations
